Rename file names in accordance with Unix way
Renames the filename given as a first argument to a new name that has
whitespaces replaced with underscores.
Also collapses multiple trailing underscores and hyphens
to a single character.

If '-d' (as in dry run) is given as a second argument, only print the new
filename without renaming the actual file.